Switch software UART over to timer 1 and remove timer prescaling to try to prevent integer rounding errors and thus minimize the possible error of the UART code.

+++ Projects/XPLAINBridge/Lib/SoftUART.h
@@ -43,7 +43,7 @@
 
 	/* Macros: */
 		#define BAUD       9600
-		#define BIT_TIME   (uint16_t)((F_CPU + (BAUD / 2.0f)) / BAUD)
+		#define BIT_TIME   ((F_CPU + (BAUD / 2)) / BAUD)
 
 		#define SRX        PD0
 		#define SRXPIN     PIND
